pyproject.toml
[project]
name = "logging_course"
version = "0.0.1"
description = "logging_course"
readme = "README.md"
license = "MIT"
authors = [{ name = "Seu nome", email = "seu@email.com" }]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]

[project.urls]
Homepage = "https://www.otaviomiranda.com.br/logging-no-python-pare-de-usar-print-no-lugar-errado/"
Documentation = "https://www.otaviomiranda.com.br/logging-no-python-pare-de-usar-print-no-lugar-errado/"
Repository = "https://github.com/luizomf/logging_course_yt.git"

[project.optional-dependencies]
dev = ["ruff", "pyright", "pytest", "pytest-xdist"]

[project.scripts]
logging_course = "logging_course.main:run"

[tool.ruff]
line-length = 88
target-version = "py313"
fix = true
show-fixes = true
indent-width = 4
exclude = ["venv", ".venv", "env", ".env", "node_modules", "__pycache__"]

[tool.ruff.lint]
select = [
    "ASYNC",
    "A",
    "ANN",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FBT",
    "FIX",
    "I",
    "ICN",
    "ISC",
    "Q",
    "RET",
    "RSE",
    "S",
    "SIM",
    "SLF",
    "T10",
    "T20",
    "TC",
    "TD",
    "TRY",
    "UP",
    "W",
    "YTT",
    "RUF",
    "N",
]
ignore = ["T201", "COM812"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
known-first-party = ["logging_course"]


# ============================
# Tipagem (Pyright)
# ============================
[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.13"
include = ["src", 'tests']
exclude = [
    "**/venv",
    "**/.venv",
    "**/env",
    "**/.env",
    "**/node_modules",
    "**/__pycache__",
]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]

[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }
